tidying

+++ cmd/imx6.go
@@ -46,9 +46,9 @@ func date(epoch int64) {
 	imx6ul.ARM.SetTimer(epoch)
 }
 
-func mem(start uint32, size int, w []byte) (b []byte) {
+func mem(start uint, size int, w []byte) (b []byte) {
 	// temporarily map page zero if required
-	if z := uint32(1 << 20); start < z {
+	if z := uint32(1 << 20); uint32(start) < z {
 		imx6ul.ARM.ConfigureMMU(0, z, (arm.TTE_AP_001<<10)|arm.TTE_SECTION)
 		defer imx6ul.ARM.ConfigureMMU(0, z, 0)
 	}
